backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.9
pydantic==2.10.4
httpx==0.27.2
google-generativeai==0.8.3
python-docx==1.1.2
python-dotenv==1.0.1
